moved util classes to plugin, spotless apply
