+++ jadx-plugins/jadx-kotlin-metadata/build.gradle.kts
@@ -4,4 +4,7 @@ plugins {
 
 dependencies {
 	api(project(":jadx-core"))
+
+	implementation("org.jetbrains.kotlinx:kotlinx-metadata-jvm:0.6.0")
+	implementation("org.jetbrains.kotlin:kotlin-compiler-embeddable:1.8.21")
 }
