+++ jadx-plugins/jadx-kotlin-metadata/src/main/kotlin/jadx/plugins/kotlin/metadata/model/KotlinMetadataConsts.kt
@@ -1,0 +1,12 @@
+package jadx.plugins.kotlin.metadata.model
+
+object KotlinMetadataConsts {
+	const val KOTLIN_METADATA_ANNOTATION = "Lkotlin/Metadata;"
+	const val KOTLIN_METADATA_K_PARAMETER = "k"
+	const val KOTLIN_METADATA_D1_PARAMETER = "d1"
+	const val KOTLIN_METADATA_D2_PARAMETER = "d2"
+	const val KOTLIN_METADATA_MV_PARAMETER = "mv"
+	const val KOTLIN_METADATA_XS_PARAMETER = "xs"
+	const val KOTLIN_METADATA_PN_PARAMETER = "pn"
+	const val KOTLIN_METADATA_XI_PARAMETER = "xi"
+}
